impruve architype and code

+++ src/main/java/com/example/avito/dto/review/ReviewDto.java
@@ -2,10 +2,12 @@ package com.example.avito.dto.review;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class ReviewDto {
     private String username;
     private String type;
-    private Byte rating;
+    private BigDecimal rating;
     private String comment;
 }
